rust/ffi: move conf_get helper to ffi crate
Ticket: 7666
