+++ rust/ffi/src/lib.rs
@@ -15,6 +15,7 @@
  * 02110-1301, USA.
  */
 
+pub mod conf;
 pub mod debug;
 pub mod detect;
 pub mod eve;
